+++ src/main/java/net/silentchaos512/gear/data/loot/ModLootModifierProvider.java
@@ -13,9 +13,7 @@ import net.neoforged.neoforge.data.event.GatherDataEvent;
 import net.silentchaos512.gear.SilentGear;
 import net.silentchaos512.gear.loot.condition.HasTraitCondition;
 import net.silentchaos512.gear.loot.modifier.BonusDropsTraitLootModifier;
-import net.silentchaos512.gear.loot.modifier.FortuneTraitLootModifier;
 import net.silentchaos512.gear.loot.modifier.MagmaticTraitLootModifier;
-import net.silentchaos512.gear.loot.modifier.SilkTouchTraitLootModifier;
 import net.silentchaos512.gear.setup.SgLoot;
 import net.silentchaos512.gear.util.Const;
 
@@ -37,45 +35,6 @@ public class ModLootModifierProvider extends GlobalLootModifierProvider {
                 new MagmaticTraitLootModifier(
                         new LootItemCondition[]{
                                 HasTraitCondition.builder(Const.Traits.MAGMATIC).build()
-                        }
-                )
-        );
-
-        add(
-                "silk_touch_trait",
-                new SilkTouchTraitLootModifier(
-                        new LootItemCondition[] {
-                                HasTraitCondition.builder(Const.Traits.SILKY).build()
-                        }
-                )
-        );
-
-        add(
-                "fortune_trait_1",
-                new FortuneTraitLootModifier(
-                        1,
-                        new LootItemCondition[] {
-                                HasTraitCondition.builder(Const.Traits.FORTUNATE, 1, 1).build()
-                        }
-                )
-        );
-
-        add(
-                "fortune_trait_2",
-                new FortuneTraitLootModifier(
-                        2,
-                        new LootItemCondition[] {
-                                HasTraitCondition.builder(Const.Traits.FORTUNATE, 2, 2).build()
-                        }
-                )
-        );
-
-        add(
-                "fortune_trait_3",
-                new FortuneTraitLootModifier(
-                        3,
-                        new LootItemCondition[] {
-                                HasTraitCondition.builder(Const.Traits.FORTUNATE, 3, 3).build()
                         }
                 )
         );