+++ src/main/java/net/silentchaos512/gear/data/trait/TraitsProvider.java
@@ -1,5 +1,6 @@
 package net.silentchaos512.gear.data.trait;
 
+import com.google.common.collect.ImmutableMap;
 import net.minecraft.core.component.DataComponentPatch;
 import net.minecraft.core.component.DataComponents;
 import net.minecraft.data.DataGenerator;
@@ -13,6 +14,7 @@ import net.minecraft.world.food.FoodProperties;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.item.crafting.Ingredient;
+import net.minecraft.world.item.enchantment.Enchantments;
 import net.minecraft.world.level.block.Blocks;
 import net.neoforged.neoforge.common.NeoForgeMod;
 import net.neoforged.neoforge.common.Tags;
@@ -253,20 +255,29 @@ public class TraitsProvider extends TraitsProviderBase {
 
         // Enchantment
 
-        // FIXME: Bring back enchantment trait
-        ret.add(new TraitBuilder(Const.Traits.FIERY, 2));
-        ret.add(new TraitBuilder(Const.Traits.SILKY, 1));
-        ret.add(new TraitBuilder(Const.Traits.FORTUNATE, 3));
-
-        /*ret.add(new EnchantmentTraitBuilder(Const.Traits.FIERY, 2)
-                .addEnchantments(GearTypes.MELEE_WEAPON, Enchantments.FIRE_ASPECT, 1, 2)
-                .addEnchantments(GearTypes.RANGED_WEAPON, Enchantments.FLAMING_ARROWS, 1)
+        ret.add(new TraitBuilder(Const.Traits.FIERY, 2)
                 .withGearTypeCondition(GearTypes.WEAPON)
+                .effects(
+                        new EnchantmentTraitEffect(
+                                ImmutableMap.of(
+                                        Enchantments.FIRE_ASPECT, List.of(1, 2),
+                                        Enchantments.FLAME, List.of(1)
+                                )
+                        )
+                )
         );
-        ret.add(new EnchantmentTraitBuilder(Const.Traits.SILKY, 1)
-                .addEnchantments(GearTypes.HARVEST_TOOL, Enchantments.SILK_TOUCH, 1)
+        ret.add(new TraitBuilder(Const.Traits.SILKY, 1)
                 .withGearTypeCondition(GearTypes.HARVEST_TOOL)
-        );*/
+                .effects(
+                        EnchantmentTraitEffect.single(Enchantments.SILK_TOUCH, 1)
+                )
+        );
+        ret.add(new TraitBuilder(Const.Traits.FORTUNATE, 3)
+                .withGearTypeCondition(GearTypes.HARVEST_TOOL)
+                .effects(
+                        EnchantmentTraitEffect.single(Enchantments.FORTUNE, 1, 2, 3)
+                )
+        );
 
         // Wielder Effect (Potion)